requirements.txt
rank_bm25
rank-llm==0.12.8
chromadb
langchain
langchain-core
langchain-community
langchain-openai
chromadb 
langchain_cohere
docling
huggingface_hub[hf_xet]
streamlit
langgraph
pyyaml
python-dotenv